Moved input boxes to separate file
