+++ src/command-manger.ts
@@ -1,54 +1,8 @@
-import { InputBoxOptions, window } from "vscode";
+import { window } from "vscode";
 import { Command } from "./command";
 import { updateConfiguration } from "./configuration";
 import { DEFAULT_PATH } from "./constants";
-
-/**
- * Return a custom input box for inserting a new command.
- * @param title The title of the input box.
- * @param prompt The text to display underneath the input box.
- * @param placeHolder A placeholder in the input box.
- * @param value The value to pre-fill in the input box.
- * @returns
- */
-function inputBox(
-  title: string,
-  prompt: string,
-  placeHolder: string,
-  value?: string
-): InputBoxOptions {
-  return {
-    title: title,
-    prompt: prompt,
-    placeHolder: placeHolder,
-    value: value,
-  };
-}
-
-/**
- * Return the input box for the `command.command`.
- * @returns The input box for the `command.command`.
- */
-function commandInputBox(): InputBoxOptions {
-  return inputBox("Command", "Command", "Command");
-}
-
-/**
- * Returns the input box for the `command.name`.
- * @param defaultNameValue The default value to pre-fill in the input box.
- * @returns The input box for the `command.name`.
- */
-function nameInputBox(defaultNameValue: string): InputBoxOptions {
-  return inputBox("Name", "Name", "Name", defaultNameValue);
-}
-
-/**
- * Returns the input box for the `command.path`.
- * @returns The input box for the `command.path`.
- */
-function pathInputBox(): InputBoxOptions {
-  return inputBox("Path", "Path", "Path", DEFAULT_PATH);
-}
+import { commandInputBox, nameInputBox, pathInputBox } from "./input-boxes";
 
 /**
  * Create a new command.
